+++ code-with-quarkus/src/main/java/org/acme/PaymentDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,5 +14,5 @@ import lombok.Setter;
 public class PaymentDTO {
     String token;
     String mid;
-    int amount;
+    BigDecimal amount;
 }
